added solution status alias
