+++ pyodsp/dec/bd/alg_leaf_pyomo.py
@@ -10,7 +10,7 @@ from .alg_leaf import BdAlgLeaf
 from ..utils import CouplingData, get_nonzero_coefficients_from_model
 from pyodsp.alg.cuts import Cut, OptimalityCut, FeasibilityCut
 from pyodsp.solver.pyomo_solver import PyomoSolver
-from pyodsp.alg.const import DEC_CUT_ABS_TOL
+from pyodsp.alg.params import DEC_CUT_ABS_TOL
 
 
 class BdAlgLeafPyomo(BdAlgLeaf):
