+++ pyodsp/dec/dd/run.py
@@ -2,6 +2,7 @@ from typing import List, Dict
 from pathlib import Path
 
 from pyodsp.alg.cuts import Cut, OptimalityCut, FeasibilityCut
+from pyodsp.alg.const import *
 
 from .logger import DdLogger
 from .node import DdNode
@@ -69,7 +70,7 @@ class DdRun:
         cuts_dn = self._run_leaf([0.0 for _ in range(self.root.num_constrs)])
         while True:
             status, solution = self.root.run_step(cuts_dn)
-            if status != 0:
+            if status != STATUS_NOT_FINISHED:
                 break
             cuts_dn = self._run_leaf(solution)
